python/requirements.txt
aiofiles
polars-lts-cpu
